Updated index js and added Reset js to srtip out function ran build
