+++ src/Reset.js
@@ -1,0 +1,22 @@
+
+import React from 'react'
+import Button from './Button'
+
+const Reset = () => {
+
+function resetProduct() {
+  	document.getElementById('productLogo').src='img/DPW_Skull_Logo_Mask.jpg';
+
+	document.getElementById("eyesBlue").checked = false;
+	document.getElementById("eyesGreen").checked = false;
+	document.getElementById("eyesGrey").checked = false;
+		
+}
+
+return (
+	<Button onClick={resetProduct}>Reset to original</Button>
+	);
+}
+
+
+export default Reset
