+++ src/index.js
@@ -10,17 +10,10 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import Button from './Button';
+import Reset from './Reset';
 
-const rootElement = document.getElementById('appButton')
+render(
+    <Reset />,
+    document.getElementById('appButton')
+);
 
-function resetProduct() {
-  	document.getElementById('productLogo').src='img/DPW_Skull_Logo_Mask.jpg';
-
-	document.getElementById("eyesBlue").checked = false;
-	document.getElementById("eyesGreen").checked = false;
-	document.getElementById("eyesGrey").checked = false;
-		
-}
-
-render(<Button onClick={resetProduct}>Reset to original</Button>, rootElement)
